GRVDEV-147 Renaming package resources
