--- aws/camunda/java/api/src/main/java/io/gravicore/loc/security/jwt/JwtUnauthorized.java
@@ -1,23 +1,0 @@
-package io.gravicore.loc.security.jwt;
-
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.web.AuthenticationEntryPoint;
-import org.springframework.stereotype.Component;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-import java.io.Serializable;
-
-@Component
-public class JwtUnauthorized implements AuthenticationEntryPoint, Serializable {
-
-    private static final long serialVersionUID = 3798723588865329956L;
-
-    @Override
-    public void commence(final HttpServletRequest request,
-                         final HttpServletResponse response,
-                         final AuthenticationException exception) throws IOException {
-        response.sendError(HttpServletResponse.SC_UNAUTHORIZED, "Unauthorized");
-    }
-}
