+++ aws/camunda/java/app/src/main/java/io/gravicore/app/cognito/model/CognitoGroup.java
@@ -1,0 +1,32 @@
+package io.gravicore.app.cognito.model;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class CognitoGroup {
+
+    private String id;
+    private Profile profile;
+
+    public String getRoleName() {
+        return profile.getName();
+    }
+
+    @Data
+    @AllArgsConstructor
+    @NoArgsConstructor
+    @Builder
+    @JsonIgnoreProperties(ignoreUnknown = true)
+    public static class Profile {
+        private String name;
+        private String description;
+    }
+}
