+++ aws/camunda/java/app/src/main/java/io/gravicore/app/config/CognitoConfig.java
@@ -1,0 +1,22 @@
+package io.gravicore.app.config;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import software.amazon.awssdk.regions.Region;
+import software.amazon.awssdk.services.cognitoidentityprovider.CognitoIdentityProviderClient;
+
+@Configuration
+public class CognitoConfig {
+
+    @Value("${aws.default.region}")
+    private String region;
+
+    @Bean
+    public CognitoIdentityProviderClient cognitoIdentityProviderClient() {
+        return CognitoIdentityProviderClient.builder()
+                .region(Region.of(region))
+            .build();
+    }
+
+}
