--- aws/camunda/java/app/src/main/java/io/gravicore/loc/config/CognitoConfig.java
@@ -1,22 +1,0 @@
-package io.gravicore.loc.config;
-
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import software.amazon.awssdk.regions.Region;
-import software.amazon.awssdk.services.cognitoidentityprovider.CognitoIdentityProviderClient;
-
-@Configuration
-public class CognitoConfig {
-
-    @Value("${aws.default.region}")
-    private String region;
-
-    @Bean
-    public CognitoIdentityProviderClient cognitoIdentityProviderClient() {
-        return CognitoIdentityProviderClient.builder()
-                .region(Region.of(region))
-            .build();
-    }
-
-}
