--- aws/camunda/java/app/src/main/java/io/gravicore/loc/custom/entity/CustomGroup.java
@@ -1,17 +1,0 @@
-package io.gravicore.loc.custom.entity;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import org.camunda.bpm.engine.identity.Group;
-
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class CustomGroup implements Group {
-    private String id;
-    private String name;
-    private String type;
-}
